Added Init Code & Core Design
I have added the Core Design behind the window size and location, I will be adding a control to switch between a few different feeds that are live on Youtube. In addition I need to make it so it reads from a JSON API to decide what to allow a user to select.